Working on auto setup

+++ developer/build.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ ! $(basename $(pwd))='developer' ]]; then
+    echo 'This script should run in the developer directory.'
+    exit
+fi
+
 cd ..
 rm -r build
 python setup.py develop
